[REFACTOR] 파일 구조 변경 - completed와 incompleted로 명시적 Section 분리
