+++ src/components/attendance/incompleted/component/AttendanceBox.tsx
